add: 강제 형 변환 추가

+++ src/com/inflearn/lecture/basic/memoryAndStatic/problem/MathArrayUtils.java
@@ -13,8 +13,9 @@ public class MathArrayUtils {
         return result;
     }
 
-    public static int average(int[] array) {
-        int sum = sum(array);
+    public static double average(int[] array) {
+        double sum = (double) sum(array);
+
         return sum / array.length;
     }
